Split C API to the modules corresponding to Rust API modules
